requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==5.2.1
deep-translator==1.11.4
distro==1.9.0
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-jalali==7.3.0
django-parler==2.3
django-parler-rest==2.2
django-rosetta==0.10.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
executing==2.2.0
frozenlist==1.5.0
googletrans==4.0.2
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipython==8.34.0
ipython_pygments_lexers==1.1.1
jalali_core==1.0.0
jdatetime==5.2.0
jedi==0.19.2
jiter==0.9.0
joblib==1.4.2
libretranslatepy==2.1.1
lxml==5.3.1
matplotlib-inline==0.1.7
mtranslate==1.8
multidict==6.2.0
nltk==3.9.1
numpy==2.2.4
openai==0.28.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
polib==1.2.0
prompt_toolkit==3.0.50
propcache==0.3.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-telegram-bot==22.0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
rosetta==0.3
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
translate==3.6.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
yarl==1.18.3
